modified Lexer to hold a vector of vectors of CommandsData to support chained while and ifs.
added parsePrint and parseClose.

changed bool m_readFromPlaceHolder to int, representing an index of placeHolder to read from. Also added "m_foundCurlyBrackets" to be triggered when '{' is sighted.

parseIf and parseWhile will both ask if '{' was found, and if so will call for a new placeholder and increase the place holders count.

some other modifications were made.